Revise pr#30503 for code style
